requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

bcrypt==5.0.0
cryptography==46.0.3
maskpass==0.3.7
python-dotenv==1.1.1
